requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-jose[cryptography]
cryptography
sqlalchemy
pydantic